# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZippyFixer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateCheckoutRequest {
  email: string;
}

export interface CreateCheckoutResponse {
  /** Stripe hosted checkout URL to redirect the user to */
  url: string;
  sessionId: string;
}

export interface VerifyPurchaseResponse {
  paid: boolean;
  downloadToken?: string;
  email?: string;
}

export interface ErrorResponse {
  error: string;
}