# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZippyFixer API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Create Stripe checkout session for ZippyFixer ($300)
 */
export const CreateCheckoutSessionBody = zod.object({
  "email": zod.string().email()
})

export const CreateCheckoutSessionResponse = zod.object({
  "url": zod.string().describe('Stripe hosted checkout URL to redirect the user to'),
  "sessionId": zod.string()
})


/**
 * @summary Download ZippyFixer ZIP using a valid purchase token
 */
export const DownloadToolParams = zod.object({
  "token": zod.coerce.string()
})


/**
 * @summary Verify payment and get download token
 */
export const VerifyPurchaseParams = zod.object({
  "sessionId": zod.coerce.string()
})

export const VerifyPurchaseResponse = zod.object({
  "paid": zod.boolean(),
  "downloadToken": zod.string().optional(),
  "email": zod.string().optional()
})