# lib/api-zod/src/generated/types/createCheckoutResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZippyFixer API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCheckoutResponse {
  /** Stripe hosted checkout URL to redirect the user to */
  url: string;
  sessionId: string;
}