# lib/api-zod/src/generated/types/verifyPurchaseResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZippyFixer API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyPurchaseResponse {
  paid: boolean;
  downloadToken?: string;
  email?: string;
}